new changes

+++ tgbot/handlers/echo.py
@@ -33,7 +33,8 @@ async def bot_echo_all(message: types.Message, state: FSMContext):
 
 
 async def remove_join_chat_message(m: types.Message):
-    await m.delete()
+    if m.new_chat_members:
+        await m.delete()
 
 
 # state_name = await state.get_state()
@@ -48,4 +49,5 @@ async def remove_join_chat_message(m: types.Message):
 def register_echo(dp: Dispatcher):
     dp.register_message_handler(bot_echo)
     dp.register_message_handler(bot_echo_all, state="*", content_types=types.ContentTypes.ANY)
-    dp.register_message_handler(remove_join_chat_message, content_types=['new_chat_members', 'left_chat_member'])
+    dp.register_message_handler(remove_join_chat_message, state="*",
+                                content_types=['new_chat_members', 'left_chat_member'])
